requirements.txt
aiohttp==3.9.1
aiohttp-cors==0.7.0
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
blessed==1.20.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorful==0.5.5
contourpy==1.2.0
cycler==0.12.1
distlib==0.3.7
filelock==3.13.1
fonttools==4.46.0
frozenlist==1.4.0
google-api-core==2.15.0
google-auth==2.25.2
googleapis-common-protos==1.62.0
gpustat==1.1.1
grpcio==1.60.0
idna==3.6
imageio==2.33.1
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
kiwisolver==1.4.5
lazy_loader==0.3
matplotlib==3.8.2
msgpack==1.0.7
multidict==6.0.4
networkx==3.2.1
numpy==1.26.2
nvidia-ml-py==12.535.133
opencensus==0.11.3
opencensus-context==0.1.3
packaging==23.2
pandas==2.1.3
Pillow==10.1.0
pipe==2.0
platformdirs==3.11.0
prometheus-client==0.19.0
protobuf==4.25.1
psutil==5.9.6
py-spy==0.3.14
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==1.10.13
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
ray==2.8.1
referencing==0.32.0
requests==2.31.0
rpds-py==0.13.2
rsa==4.9
scikit-image==0.22.0
scipy==1.11.4
six==1.16.0
smart-open==6.4.0
tifffile==2023.12.9
tqdm==4.66.1
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.1.0
virtualenv==20.21.0
wcwidth==0.2.12
yarl==1.9.4
